+++ CHANGELOG.md
@@ -13,6 +13,11 @@ This Changelog tracks changes to this project. The notes below include a summary
 
 - `fixed` bug in building user permissions for files on the portal
   - allowed users with cross-assay permissions to see clinical data files
+  
+## Version `0.27.8` - 14 Oct 2022
+
+- `added` credentialing to dataset update
+- `fixed` project id bug
 
 ## Version `0.27.7` - 14 Oct 2022